% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_datasets.R
\name{sim_data_mean_shift}
\alias{sim_data_mean_shift}
\title{Simulate a mean shift DID data-set}
\usage{
sim_data_mean_shift(
  nobs = 300,
  ngroups = 3,
  group_treat_order = c(1, 2, 3),
  years = c(1980, 2010),
  group_treat_time = c(1990, 2000, 9999),
  tau_vec = c(1, 1, 1),
  seed = 1
)
}
\arguments{
\item{nobs}{Number of observations}

\item{ngroups}{Number of groups}

\item{group_treat_order}{Order of group treatment. If groups have same treatment timing, then omit from argument. For example, if three groups (ngroups = 3), and groups 1 and 2 are both treated at same time, input is: 'group_treat_timing = c(1,3)'}

\item{years}{Minimum and maximum time periods}

\item{group_treat_time}{Timing each group gets treated. From above example, if inputted 'group_treat_timing = c(1,3)' -- i.e. groups 1:2 are treated first and groups 3: are treated after, the need to input two time periods, one for first treatment and one for later treatment. If later are never-treated, input a time period after maximum time period in the data, for example 'group_treat_time=c(1995, 9999)'}

\item{tau_vec}{A vector of treatment effects. For hetergenous data-sets, simply input different numbers, e.g. 'tau_vec = c(1,.5)'}

\item{seed}{A random seed}
}
\value{
A simulated data frame for DID analysis.
}
\description{
Simulate a mean shift DID data-set
}
\examples{
df <- sim_data_mean_shift()
}
